backend/requirements.txt
fastapi==0.109.1
uvicorn==0.17.4
python-dotenv==1.0.1
beanie==1.26.0
certifi==2024.6.2
passlib==1.7.4
pyjwt==2.4.0
python-multipart==0.0.6